chore(model): RPC conversion code isolation (A-D)
