+++ crates/api-model/src/controller_outcome.rs
@@ -76,41 +76,6 @@ impl From<&&'static Location<'static>> for PersistentSourceReference {
     }
 }
 
-impl From<PersistentSourceReference> for rpc::forge::ControllerStateSourceReference {
-    fn from(source_ref: PersistentSourceReference) -> Self {
-        rpc::forge::ControllerStateSourceReference {
-            file: source_ref.file,
-            line: source_ref.line.try_into().unwrap_or_default(),
-        }
-    }
-}
-
-impl From<PersistentStateHandlerOutcome> for rpc::forge::ControllerStateReason {
-    fn from(p: PersistentStateHandlerOutcome) -> rpc::forge::ControllerStateReason {
-        use rpc::forge::ControllerStateOutcome::*;
-        let (outcome, outcome_msg, source_ref) = match p {
-            PersistentStateHandlerOutcome::Wait { reason, source_ref } => {
-                (Wait, Some(reason), source_ref)
-            }
-            PersistentStateHandlerOutcome::Error { err, source_ref } => {
-                (Error, Some(err), source_ref)
-            }
-            PersistentStateHandlerOutcome::Transition { source_ref } => {
-                (Transition, None, source_ref)
-            }
-            PersistentStateHandlerOutcome::DoNothing { source_ref } => {
-                (DoNothing, None, source_ref)
-            }
-            PersistentStateHandlerOutcome::DoNothingWithDetails => (DoNothing, None, None),
-        };
-        rpc::forge::ControllerStateReason {
-            outcome: outcome.into(), // into converts it to i32
-            outcome_msg,
-            source_ref: source_ref.map(Into::into),
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
