+++ crates/api-model/src/dhcp_record.rs
@@ -47,22 +47,3 @@ pub struct DhcpRecord {
 
     pub last_invalidation_time: chrono::DateTime<chrono::Utc>,
 }
-
-impl From<DhcpRecord> for rpc::forge::DhcpRecord {
-    fn from(record: DhcpRecord) -> Self {
-        Self {
-            machine_id: record.machine_id,
-            machine_interface_id: Some(record.machine_interface_id),
-            segment_id: Some(record.segment_id),
-            subdomain_id: record.subdomain_id,
-            fqdn: record.fqdn,
-            mac_address: record.mac_address.to_string(),
-            address: record.address.to_string(),
-            mtu: record.mtu,
-            prefix: record.prefix.to_string(),
-            gateway: record.gateway.map(|gw| gw.to_string()),
-            booturl: None, // TODO(ajf): extend database, synthesize URL
-            last_invalidation_time: Some(record.last_invalidation_time.into()),
-        }
-    }
-}
