+++ crates/api-model/src/rpc_conv/allocation_type.rs
@@ -1,0 +1,42 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use crate::address_selection_strategy::AddressSelectionStrategy;
+use crate::allocation_type::{AllocationType, AssignStaticResult};
+
+impl From<AssignStaticResult> for rpc::forge::AssignStaticAddressStatus {
+    fn from(result: AssignStaticResult) -> Self {
+        match result {
+            AssignStaticResult::Assigned => rpc::forge::AssignStaticAddressStatus::Assigned,
+            AssignStaticResult::ReplacedStatic => {
+                rpc::forge::AssignStaticAddressStatus::ReplacedStatic
+            }
+            AssignStaticResult::ReplacedDhcp => rpc::forge::AssignStaticAddressStatus::ReplacedDhcp,
+        }
+    }
+}
+
+impl From<AddressSelectionStrategy> for AllocationType {
+    fn from(strategy: AddressSelectionStrategy) -> Self {
+        match strategy {
+            AddressSelectionStrategy::NextAvailableIp => AllocationType::Dhcp,
+            AddressSelectionStrategy::Automatic => AllocationType::Dhcp,
+            AddressSelectionStrategy::NextAvailablePrefix(_) => AllocationType::Dhcp,
+            AddressSelectionStrategy::StaticAddress(_) => AllocationType::Static,
+        }
+    }
+}
